adds tests for encryption and decryption with key nd date
