+++ lib/enigma.rb
@@ -70,17 +70,17 @@ include Key
   #   character_map[new_ind % 27]
   # end
 
-  def decrypt_translate(char, shift_amt)
-    new_ind = character_map.index(char) - shift_amt
-    character_map[new_ind % 27]
-  end
+  # def decrypt_translate(char, shift_amt)
+  #   new_ind = character_map.index(char) - shift_amt
+  #   character_map[new_ind % 27]
+  # end
 
-  def prepare_for_encryption(message, key = get_random_key, date_string = today)
-    shift = make_shift_hash(date_string, key)
-    message.chars.map.with_index do |char, i|
-      encrypt_translate(char, shift[i % 4])
-    end.join
-  end
+  # def prepare_for_encryption(message, key = get_random_key, date_string = today)
+  #   shift = make_shift_hash(date_string, key)
+  #   message.chars.map.with_index do |char, i|
+  #     encrypt_translate(char, shift[i % 4])
+  #   end.join
+  # end
 
   def encrypt(message, key = get_random_key, date_string = today)
     message = prepare_for_encryption(message, key, date_string)
@@ -91,12 +91,12 @@ include Key
     return_hash
   end
 
-  def prepare_for_decryption(message, key, date_string = today)
-    shift = make_shift_hash(date_string, key)
-    message.chars.map.with_index do |char, i|
-      decrypt_translate(char, shift[i % 4])
-    end.join
-  end
+  # def prepare_for_decryption(message, key, date_string = today)
+  #   shift = make_shift_hash(date_string, key)
+  #   message.chars.map.with_index do |char, i|
+  #     decrypt_translate(char, shift[i % 4])
+  #   end.join
+  # end
 
   def decrypt(message, key, date_string = today)
     # key = @key.get_random_key if key == nil
