+++ test/decrypt_test.rb
@@ -9,11 +9,20 @@ class DecryptTest < Minitest::Test
     assert_instance_of Decrypt, decrypt
   end
 
-  def test_it_can_prepare_for_encrpyt
+  def test_it_can_prepare_for_decrpyt
     decrypt = Decrypt.new
-    expected = 'keder ohulw'
-    actual = decrypt.prepare_for_encryption("hello world", "02715", "040895")
+    expected = 'hello world'
+    actual = decrypt.prepare_for_decryption("keder ohulw", "02715", "040895")
     assert_equal expected, actual
   end
 
+  def test_it_can_decrypt_a_string
+    actual = @enigma.decrypt("keder ohulw", "02715", "040895")
+    expected = {
+        decryption: "hello world",
+        key: "02715",
+        date: "040895"
+      }
+    assert_equal expected, actual
+  end
 end
