+++ test/encrypt_test.rb
@@ -16,4 +16,13 @@ class EncryptTest < Minitest::Test
     assert_equal expected, actual
   end
 
+  def test_it_can_encrypt_a_message_with_key_and_date
+    actual = @enigma.encrypt("hello world", "02715", "040895")
+    expected = {
+                encryption: "keder ohulw",
+                key: "02715",
+                date: "040895"
+              }
+    assert_equal expected, actual
+  end
 end
